Added more methods to Matrix class
Added new constructors for both the Matrix and ComplexNumber classes to create a new object of that specific type from another pre-existent object. Basically you can create copies of an object now. Also changed the show method of the Matrix class and added the static add method to add to Matrices together.